update the send group message flow

+++ ios/Message/TXIMMessageInfo.h
@@ -30,7 +30,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 @property(nonatomic, assign) TXIMMessageType msgType;
 
-@property(nonatomic, assign) double msgTime;
+@property(nonatomic, assign) NSInteger msgTime;
 
 @property(nonatomic, assign) BOOL isRead;
